Move test_stdlibs to file, note qsort failure
