+++ tests/core/test_stdlibs.out
@@ -1,0 +1,15 @@
+*1,2,3,5,5,6*
+*stdin==0:0*
+*%*
+*5*
+*66.0*
+*10*
+*0*
+*-10*
+*18*
+*10*
+*0*
+*4294967286*
+*malloc(0)!=0:1*
+tolower_l: a
+*cleaned*
